using DynamicDS.set_state to push events State management modified in PyPLC to avoid State set by qualities explicitly added all DynamicDS properties to PyPLC Added DynamicCommands KeepAttributes,KeepTime,CheckDependencies will have its own default values for PyPLC
Static/Mapping attributes not require polling to be up to date
--------------------------------------------------------------

In the PyPLC the KeepAttributes flag is always valid for mapped attributes.

CheckDependencies and KeepAttributes have been improved to save times.

The ReadMap processing has been reduced to <1ms; and for every Mapped array it will update dyn_values and _locals dictionary. So it is not needed to call evalAttr or read_dyn_attr to update the Mappings.

For a beamline like Circe it meant an overhead of 47 ms every second; which is still too much and may cause polling cycle exceptions!

Triggering read() in dependencies is chosen as solution:

 - Now if the last value stored for a dependency is older than KeepTime the read_dyn_attr(dependency) method will be called to update this value.

 - In addition the StateMachine() is executed every STATE_MACHINE_PERIOD (KeepTime) and will try to update DigitalsREAD if LastHWUpdate was more than a period ago.

How MapDict,MapCommands,MapFlags and ReadMap work
-------------------------------------------------

 * **MapDict** will keep *AttrName:Formula* as declared in properties.

 * **MapCommands** will contain the list of modbus commands for each mapping (result of **GetCommands4Map(formula)** call).

 * **MapFlags** [Map] is set to True when registers belonging to a Map have been updated from hardware ( **sendModbusCall().checkMaps()** call).

 * ReadMap(Attr / Regs) will process the reading of the registers associated to a Map and will store the readings in *dyn_values* and *_locals* dictionaries. It will also set the **MapFlag** to False.