Major refactor of TwickenhamSMC device server. The refactor was made mainly to overcome 2 problems:
1) High responsive device. Now the device has to be highly responsive.
Mainly, it has to respond to high rate of state requests, but it also
has to avoid getting hung for more than tango timeout default time (3000
ms). To achieve this, the communications have been separated in a new
threaded class, which also includes a method to continuously update the
device state. This method also takes into account if any other command
or communication is in progress, and gives them high priority in order
to avoid tango timeouts on commands/attributes. The reading timeout of
the serial device has also been lowered to 1500 ms, since we detected
that sometimes there are sporuous communication errors which led the
device to fault (this is related with point 2).

2) A possible hardware bug with communications (see below for more
details).

Things to note:

1) Communicator new threaded class for communications and state update.

2) self.preserver_moving is now used to preserve MOVING state when a
ramp has just been started.

3) "CommErrors" new attribute. A problem has been detected with the
communications with the hardware (see test directory for more details).
To avoid setting FAULT the device due to a temporal communication
problem the device now simply ignore a single error, but will go to
FAULT if there is more than one consecutive error. Furthermore, the
communications problems are now logged (max 100 errors).